Adds unfinished scenes for rounds 1 and 2
These scenes still need transitions between each other, which will be done by placing enemies in each scene in a group for enemies, and obtaining the amount of enemies at the start of each round. The enemies will have signals for death that the main_gameplay scene is listening to, decrementing the enemy count until it's zero to transition to the next round.